Create Engblog::ArticleImport
Smallest TDD step I can take towards this story. Choosing to place this
in lib/engblog because this doesn't really fit into one of Rails
predefined folders. Chose `Engblog` because that's the namespace the app
was created with as shown in `application.rb`. Note the use of YARD
class documentation.